Refactored tests all into the API folder
